renamed processors ops and utils
fixed examples
