+++ Wrappers/Python/conda-recipe/meta.yaml
@@ -19,7 +19,7 @@ requirements:
     - python
     - numpy
     - scipy
-    - ccpi-common
+    - ccpi-framework
     - astra-toolbox
 	
 about:
