+++ Wrappers/Python/wip/demo_sophiabeads.py
@@ -3,8 +3,8 @@ from ccpi.io.reader import XTEKReader
 import numpy as np
 import matplotlib.pyplot as plt
 from ccpi.framework import ImageGeometry, AcquisitionGeometry, AcquisitionData, ImageData
-from ccpi.astra.astra_ops import AstraProjectorSimple
-from ccpi.reconstruction.algs import CGLS
+from ccpi.astra.ops import AstraProjectorSimple
+from ccpi.optimisation.algs import CGLS
 
 # Set up reader object and read the data
 datareader = XTEKReader("C:/Users/mbbssjj2/Documents/SophiaBeads_256_averaged/SophiaBeads_256_averaged.xtekct")
